Добавлена анимация взрыва мусора
